Fix critical typo: "gteHardFork(...)" -> "getHardFork(...)"
NOT A COMPLETE RESOLUTION. Calling the code in question results in an error: "TypeError: this._common.gteHardfork is not a function". What's interesting is this actually SHOULD work, since the method is spelled the same on the Common class. However with this error coming from an abstracted layer of code I'm interacting with, I believe this is an issue with the code, and not user error. Anyways here's the version with all the "typos" corrected.